requirements.txt
gTTS==2.5
openai==1.13.3
SpeechRecognition==3.10.1
PyAudio==0.2.11
playsound==1.3.0
pyyaml==6.0.1
plac==1.4.3
piney-event==0.0.2
pygobject==3.46.0
